fix: some types
