+++ src/navigation/navigator.ts
@@ -2,23 +2,7 @@ import { bindActionCreators, Store } from "redux";
 
 import { actionCreators as navigationActionCreators } from "@state/navigation/actions";
 
-export interface NavigatorParams {
-  [key: string]: unknown;
-}
-
-export interface NavigatorT {
-  initialize: (store: Store, screens: unknown) => this;
-  store: Store | null;
-  navigate: (
-    screenName: string,
-    params?: NavigatorParams
-  ) => void | (() => void);
-  push: (screenName: string, params?: NavigatorParams) => void | (() => void);
-}
-
-// WithNavigatorT - will always return the store, omit & attach to generic
-export type WithNavigatorT<T = any> = T &
-  Omit<NavigatorT, "initialize" | "store"> & { store: Store };
+import { NavigatorT } from "./types";
 
 export const registeredScreens = {
   Search: () => require("./search"),
